refactor: change app package name
