+++ app/src/main/java/com/android/voyageur/resources/C.kt
@@ -1,4 +1,4 @@
-package com.android.sample.resources
+package com.android.voyageur.resources
 
 // Like R, but C
 object C {
