+++ app/src/main/java/com/android/voyageur/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.android.voyageur.ui.theme
 
 import androidx.compose.ui.graphics.Color
