+++ app/src/main/java/com/android/voyageur/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.android.voyageur.ui.theme
 
 import android.app.Activity
 import android.os.Build
