+++ app/src/test/java/com/android/voyageur/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.android.voyageur
 
 import org.junit.Assert.*
 import org.junit.Test
