+++ app/src/test/java/com/android/voyageur/PointTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.android.voyageur
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
